Refactor preprocessing.
